assert_has_dims {assertive.properties}	R Documentation

Does the input have dimensions?

Description

Checks to see if the input has dimensions.

Usage

assert_has_dims(x, severity = getOption("assertive.severity", "stop"))

has_dims(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

has_dims returnsTRUE if dim is non-null. assert_has_dims returns nothing but throws an error if has_dims is not TRUE.

See Also

dim, is_of_dimension.

[Package assertive.properties version 0.0-4 Index]