assert_has_duplicates {assertive.properties}	R Documentation

Does the input have duplicates?

Description

Checks to see if the input has duplicates.

Usage

assert_has_duplicates(x, severity = getOption("assertive.severity", "stop"))

assert_has_no_duplicates(x, severity = getOption("assertive.severity",
  "stop"))

has_duplicates(x, .xname = get_name_in_parent(x))

has_no_duplicates(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

has_duplicates returns TRUE ifanyDuplicated is TRUE. assert_has_duplicates returns nothing but throws an error if has_duplicates is not TRUE. has_no_duplicates is the negation of has_duplicates.

See Also

anyDuplicated.

Examples

x <- sample(10, 100, replace = TRUE)
assert_has_duplicates(x)
has_no_duplicates(x)

[Package assertive.properties version 0.0-4 Index]